Break up AST tree
